I hate K&R-style and lack of whitespace with a fiery passion, but I stuck with it because I figured it'd mean less cleanup for you ;)
If a warp is located inside of non-air blocks, players who warp there will suffocate. This fix forces the warp to "jump upward" until two free blocks of air are found, and only then, the player is warped.

Note that the bottom player.teleport() line preserves the yaw and pitch from the original location. The fix has been tested and works on CB #1000.